#52 [style] bookmark, star 디렉토리의 command 디렉토리 제거
구분 지어 사용할 이유가 없기에 command 디렉토리만 삭제합니다. 코드나 프로젝트의 구조를 큰 변화를 주지 않기에 style로 간주했습니다!
